requirements.txt
gunicorn
dash
plotly
pandas
dash_bootstrap_components
azure-storage-blob
dash-ag-grid
djangorestframework
django_plotly_dash
dj-database-url
whitenoise
pillow
